code/Retrieval/services/reranking/v1.0.0/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
torch>=2.0.0
